refactor(config): centralize config values & move helper funcs
move all game config values to config.py and move helper functions for
game to utils.py
