+++ woodoku/entity/score_agent.py
@@ -1,6 +1,4 @@
-GROUP_POINTS = 18
-STREAK_POINTS = 10
-COMBO_POINTS = 28
+from config import COMBO_POINTS, GROUP_POINTS, STREAK_POINTS
 
 
 class ScoreAgent:
